requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.43.0
datasets>=2.19.0
faiss-cpu>=1.7.4
bitsandbytes>=0.43.0 ; platform_system != 'Windows'
peft>=0.11.0
accelerate>=0.33.0
trl>=0.9.6
PyYAML>=6.0.1
Pillow>=10.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
rich>=13.7.0

